DAOS-10368 test: Don't use t.Skip() for disabled tests
These tests show up in coverage reports and have to be
revisited on every release. The tests are unable to run
in CI for various reasons, but are still valuable to
run locally in developer environments.

Removes an unreliable and overly-complicated agent
integration test that duplicates unit test coverage.


Required-githooks: true
