+++ src/control/lib/hardware/ucx/ucx_test.go
@@ -7,10 +7,7 @@
 package ucx
 
 import (
-	"context"
-	"fmt"
 	"testing"
-	"time"
 
 	"github.com/google/go-cmp/cmp"
 
@@ -18,32 +15,6 @@ import (
 	"github.com/daos-stack/daos/src/control/common/test"
 	"github.com/daos-stack/daos/src/control/logging"
 )
-
-func TestUCX_Provider_GetFabricInterfaces_Integrated(t *testing.T) {
-	cleanup, err := Load()
-	if err != nil {
-		t.Skipf("can't load lib (%s)", err.Error())
-	}
-	defer cleanup()
-
-	// Can't mock the underlying UCX calls, but we can make sure it doesn't crash or
-	// error on the normal happy path.
-
-	log, buf := logging.NewTestLogger(t.Name())
-	defer test.ShowBufferOnFailure(t, buf)
-
-	p := NewProvider(log)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	result, err := p.GetFabricInterfaces(ctx)
-
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-
-	fmt.Printf("FabricInterfaceSet:\n%s\n", result)
-}
 
 func TestUCX_Provider_getProviderSet(t *testing.T) {
 	for name, tc := range map[string]struct {
